refactor: extract initializeStandalone into shared module
standalone.js and generateEntryPointSource now both import from
src/initializeStandalone.js, eliminating the duplicated function
body that could drift between the two.
